+++ src/components/Header/__tests__/Header.test.tsx
@@ -1,31 +1,31 @@
 import { fireEvent, render, screen } from '@testing-library/react';
 
-import Header from '..';
+import { Header } from '../Header';
 
 describe('Header', () => {
   it('should render header', () => {
     render(<Header title="Header" />);
 
-    expect(screen.getByRole('banner', { name: 'Header' })).toBeVisible();
+    expect(screen.getByRole('banner')).toBeVisible();
   });
 
   it('should render back button in header', () => {
     render(<Header title="Header" onGoBackRequest={jest.fn()} />);
 
-    expect(screen.getByRole('button', { name: 'Go Back' })).toBeVisible();
+    expect(screen.getByRole('button', { name: 'Go back' })).toBeVisible();
   });
 
   it('should not render back button in header', () => {
     render(<Header title="Header" />);
 
-    expect(screen.queryByRole('button', { name: 'Go Back' })).not.toBeInTheDocument();
+    expect(screen.queryByRole('button', { name: 'Go back' })).not.toBeInTheDocument();
   });
 
   it('should navigate back when back button is clicked', () => {
     const onGoBackRequest = jest.fn();
     render(<Header title="Header" onGoBackRequest={onGoBackRequest} />);
 
-    fireEvent.click(screen.getByRole('button', { name: 'Go Back' }));
+    fireEvent.click(screen.getByRole('button', { name: 'Go back' }));
 
     expect(onGoBackRequest).toHaveBeenCalled();
   });
